fix: 一些小bug

+++ DoubanFM/Model/SongInfo.m
@@ -40,7 +40,7 @@ static int currentSongIndex;
     currentSong = songInfo;
 }
 
-+ (int)currentSongIndex{
++ (NSInteger)currentSongIndex{
     return currentSongIndex;
 }
